CryptoAlg-858: Bring EVP_DigestSign/Verify APIs within fipsmodule
Call-out:
- evp_test.cc tests both approved and non-approved algorithms so it
was not moved under fipsmodule (for now, at least).
- EVP_Sign/Verify APIs were left outside fipsmodule since they are
designated as "old functions".
