+++ crypto/evp_extra/p_ec_asn1.c
@@ -9,7 +9,7 @@
  * are met:
  *
  * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer. 
+ *    notice, this list of conditions and the following disclaimer.
  *
  * 2. Redistributions in binary form must reproduce the above copyright
  *    notice, this list of conditions and the following disclaimer in
@@ -62,7 +62,7 @@
 #include <openssl/ecdsa.h>
 #include <openssl/err.h>
 
-#include "internal.h"
+#include "../fipsmodule/evp/internal.h"
 
 
 static int eckey_pub_encode(CBB *out, const EVP_PKEY *key) {
