+++ crypto/evp_extra/p_ed25519.c
@@ -18,7 +18,7 @@
 #include <openssl/err.h>
 #include <openssl/mem.h>
 
-#include "internal.h"
+#include "../fipsmodule/evp/internal.h"
 
 
 // Ed25519 has no parameters to copy.
