+++ crypto/evp_extra/p_ed25519_asn1.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019, Google Inc.
+/* Copyright (c) 2017, Google Inc.
  *
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -19,59 +19,61 @@
 #include <openssl/err.h>
 #include <openssl/mem.h>
 
-#include "internal.h"
+#include "../fipsmodule/evp/internal.h"
 #include "../internal.h"
 
 
-static void x25519_free(EVP_PKEY *pkey) {
+static void ed25519_free(EVP_PKEY *pkey) {
   OPENSSL_free(pkey->pkey.ptr);
   pkey->pkey.ptr = NULL;
 }
 
-static int x25519_set_priv_raw(EVP_PKEY *pkey, const uint8_t *in, size_t len) {
+static int ed25519_set_priv_raw(EVP_PKEY *pkey, const uint8_t *in, size_t len) {
   if (len != 32) {
     OPENSSL_PUT_ERROR(EVP, EVP_R_DECODE_ERROR);
     return 0;
   }
 
-  X25519_KEY *key = OPENSSL_malloc(sizeof(X25519_KEY));
+  ED25519_KEY *key = OPENSSL_malloc(sizeof(ED25519_KEY));
   if (key == NULL) {
     OPENSSL_PUT_ERROR(EVP, ERR_R_MALLOC_FAILURE);
     return 0;
   }
 
-  OPENSSL_memcpy(key->priv, in, 32);
-  X25519_public_from_private(key->pub, key->priv);
+  // The RFC 8032 encoding stores only the 32-byte seed, so we must recover the
+  // full representation which we use from it.
+  uint8_t pubkey_unused[32];
+  ED25519_keypair_from_seed(pubkey_unused, key->key.priv, in);
   key->has_private = 1;
 
-  x25519_free(pkey);
+  ed25519_free(pkey);
   pkey->pkey.ptr = key;
   return 1;
 }
 
-static int x25519_set_pub_raw(EVP_PKEY *pkey, const uint8_t *in, size_t len) {
+static int ed25519_set_pub_raw(EVP_PKEY *pkey, const uint8_t *in, size_t len) {
   if (len != 32) {
     OPENSSL_PUT_ERROR(EVP, EVP_R_DECODE_ERROR);
     return 0;
   }
 
-  X25519_KEY *key = OPENSSL_malloc(sizeof(X25519_KEY));
+  ED25519_KEY *key = OPENSSL_malloc(sizeof(ED25519_KEY));
   if (key == NULL) {
     OPENSSL_PUT_ERROR(EVP, ERR_R_MALLOC_FAILURE);
     return 0;
   }
 
-  OPENSSL_memcpy(key->pub, in, 32);
+  OPENSSL_memcpy(key->key.pub.value, in, 32);
   key->has_private = 0;
 
-  x25519_free(pkey);
+  ed25519_free(pkey);
   pkey->pkey.ptr = key;
   return 1;
 }
 
-static int x25519_get_priv_raw(const EVP_PKEY *pkey, uint8_t *out,
+static int ed25519_get_priv_raw(const EVP_PKEY *pkey, uint8_t *out,
                                 size_t *out_len) {
-  const X25519_KEY *key = pkey->pkey.ptr;
+  const ED25519_KEY *key = pkey->pkey.ptr;
   if (!key->has_private) {
     OPENSSL_PUT_ERROR(EVP, EVP_R_NOT_A_PRIVATE_KEY);
     return 0;
@@ -87,14 +89,15 @@ static int x25519_get_priv_raw(const EVP_PKEY *pkey, uint8_t *out,
     return 0;
   }
 
-  OPENSSL_memcpy(out, key->priv, 32);
+  // The raw private key format is the first 32 bytes of the private key.
+  OPENSSL_memcpy(out, key->key.priv, 32);
   *out_len = 32;
   return 1;
 }
 
-static int x25519_get_pub_raw(const EVP_PKEY *pkey, uint8_t *out,
+static int ed25519_get_pub_raw(const EVP_PKEY *pkey, uint8_t *out,
                                size_t *out_len) {
-  const X25519_KEY *key = pkey->pkey.ptr;
+  const ED25519_KEY *key = pkey->pkey.ptr;
   if (out == NULL) {
     *out_len = 32;
     return 1;
@@ -105,12 +108,12 @@ static int x25519_get_pub_raw(const EVP_PKEY *pkey, uint8_t *out,
     return 0;
   }
 
-  OPENSSL_memcpy(out, key->pub, 32);
+  OPENSSL_memcpy(out, key->key.pub.value, 32);
   *out_len = 32;
   return 1;
 }
 
-static int x25519_pub_decode(EVP_PKEY *out, CBS *params, CBS *key) {
+static int ed25519_pub_decode(EVP_PKEY *out, CBS *params, CBS *key) {
   // See RFC 8410, section 4.
 
   // The parameters must be omitted. Public keys have length 32.
@@ -119,21 +122,21 @@ static int x25519_pub_decode(EVP_PKEY *out, CBS *params, CBS *key) {
     return 0;
   }
 
-  return x25519_set_pub_raw(out, CBS_data(key), CBS_len(key));
+  return ed25519_set_pub_raw(out, CBS_data(key), CBS_len(key));
 }
 
-static int x25519_pub_encode(CBB *out, const EVP_PKEY *pkey) {
-  const X25519_KEY *key = pkey->pkey.ptr;
+static int ed25519_pub_encode(CBB *out, const EVP_PKEY *pkey) {
+  const ED25519_KEY *key = pkey->pkey.ptr;
 
   // See RFC 8410, section 4.
   CBB spki, algorithm, oid, key_bitstring;
   if (!CBB_add_asn1(out, &spki, CBS_ASN1_SEQUENCE) ||
       !CBB_add_asn1(&spki, &algorithm, CBS_ASN1_SEQUENCE) ||
       !CBB_add_asn1(&algorithm, &oid, CBS_ASN1_OBJECT) ||
-      !CBB_add_bytes(&oid, x25519_asn1_meth.oid, x25519_asn1_meth.oid_len) ||
+      !CBB_add_bytes(&oid, ed25519_asn1_meth.oid, ed25519_asn1_meth.oid_len) ||
       !CBB_add_asn1(&spki, &key_bitstring, CBS_ASN1_BITSTRING) ||
       !CBB_add_u8(&key_bitstring, 0 /* padding */) ||
-      !CBB_add_bytes(&key_bitstring, key->pub, 32) ||
+      !CBB_add_bytes(&key_bitstring, key->key.pub.value, 32) ||
       !CBB_flush(out)) {
     OPENSSL_PUT_ERROR(EVP, EVP_R_ENCODE_ERROR);
     return 0;
@@ -142,13 +145,13 @@ static int x25519_pub_encode(CBB *out, const EVP_PKEY *pkey) {
   return 1;
 }
 
-static int x25519_pub_cmp(const EVP_PKEY *a, const EVP_PKEY *b) {
-  const X25519_KEY *a_key = a->pkey.ptr;
-  const X25519_KEY *b_key = b->pkey.ptr;
-  return OPENSSL_memcmp(a_key->pub, b_key->pub, 32) == 0;
+static int ed25519_pub_cmp(const EVP_PKEY *a, const EVP_PKEY *b) {
+  const ED25519_KEY *a_key = a->pkey.ptr;
+  const ED25519_KEY *b_key = b->pkey.ptr;
+  return OPENSSL_memcmp(a_key->key.pub.value, b_key->key.pub.value, 32) == 0;
 }
 
-static int x25519_priv_decode(EVP_PKEY *out, CBS *params, CBS *key) {
+static int ed25519_priv_decode(EVP_PKEY *out, CBS *params, CBS *key) {
   // See RFC 8410, section 7.
 
   // Parameters must be empty. The key is a 32-byte value wrapped in an extra
@@ -161,11 +164,11 @@ static int x25519_priv_decode(EVP_PKEY *out, CBS *params, CBS *key) {
     return 0;
   }
 
-  return x25519_set_priv_raw(out, CBS_data(&inner), CBS_len(&inner));
+  return ed25519_set_priv_raw(out, CBS_data(&inner), CBS_len(&inner));
 }
 
-static int x25519_priv_encode(CBB *out, const EVP_PKEY *pkey) {
-  X25519_KEY *key = pkey->pkey.ptr;
+static int ed25519_priv_encode(CBB *out, const EVP_PKEY *pkey) {
+  ED25519_KEY *key = pkey->pkey.ptr;
   if (!key->has_private) {
     OPENSSL_PUT_ERROR(EVP, EVP_R_NOT_A_PRIVATE_KEY);
     return 0;
@@ -177,12 +180,12 @@ static int x25519_priv_encode(CBB *out, const EVP_PKEY *pkey) {
       !CBB_add_asn1_uint64(&pkcs8, 0 /* version */) ||
       !CBB_add_asn1(&pkcs8, &algorithm, CBS_ASN1_SEQUENCE) ||
       !CBB_add_asn1(&algorithm, &oid, CBS_ASN1_OBJECT) ||
-      !CBB_add_bytes(&oid, x25519_asn1_meth.oid, x25519_asn1_meth.oid_len) ||
+      !CBB_add_bytes(&oid, ed25519_asn1_meth.oid, ed25519_asn1_meth.oid_len) ||
       !CBB_add_asn1(&pkcs8, &private_key, CBS_ASN1_OCTETSTRING) ||
       !CBB_add_asn1(&private_key, &inner, CBS_ASN1_OCTETSTRING) ||
       // The PKCS#8 encoding stores only the 32-byte seed which is the first 32
       // bytes of the private key.
-      !CBB_add_bytes(&inner, key->priv, 32) ||
+      !CBB_add_bytes(&inner, key->key.priv, 32) ||
       !CBB_flush(out)) {
     OPENSSL_PUT_ERROR(EVP, EVP_R_ENCODE_ERROR);
     return 0;
@@ -191,28 +194,28 @@ static int x25519_priv_encode(CBB *out, const EVP_PKEY *pkey) {
   return 1;
 }
 
-static int x25519_size(const EVP_PKEY *pkey) { return 32; }
+static int ed25519_size(const EVP_PKEY *pkey) { return 64; }
 
-static int x25519_bits(const EVP_PKEY *pkey) { return 253; }
+static int ed25519_bits(const EVP_PKEY *pkey) { return 253; }
 
-const EVP_PKEY_ASN1_METHOD x25519_asn1_meth = {
-    EVP_PKEY_X25519,
-    {0x2b, 0x65, 0x6e},
+const EVP_PKEY_ASN1_METHOD ed25519_asn1_meth = {
+    EVP_PKEY_ED25519,
+    {0x2b, 0x65, 0x70},
     3,
-    x25519_pub_decode,
-    x25519_pub_encode,
-    x25519_pub_cmp,
-    x25519_priv_decode,
-    x25519_priv_encode,
-    x25519_set_priv_raw,
-    x25519_set_pub_raw,
-    x25519_get_priv_raw,
-    x25519_get_pub_raw,
+    ed25519_pub_decode,
+    ed25519_pub_encode,
+    ed25519_pub_cmp,
+    ed25519_priv_decode,
+    ed25519_priv_encode,
+    ed25519_set_priv_raw,
+    ed25519_set_pub_raw,
+    ed25519_get_priv_raw,
+    ed25519_get_pub_raw,
     NULL /* pkey_opaque */,
-    x25519_size,
-    x25519_bits,
+    ed25519_size,
+    ed25519_bits,
     NULL /* param_missing */,
     NULL /* param_copy */,
     NULL /* param_cmp */,
-    x25519_free,
+    ed25519_free,
 };
