+++ crypto/fipsmodule/bcm.c
@@ -81,6 +81,11 @@
 #include "ec/simple_mul.c"
 #include "ec/util.c"
 #include "ec/wnaf.c"
+#include "evp/digestsign.c"
+#include "evp/evp.c"
+#include "evp/evp_ctx.c"
+#include "evp/p_ec.c"
+#include "evp/p_rsa.c"
 #include "hmac/hmac.c"
 #include "md4/md4.c"
 #include "md5/md5.c"
