+++ crypto/fipsmodule/evp/evp_ctx.c
@@ -62,7 +62,7 @@
 #include <openssl/err.h>
 #include <openssl/mem.h>
 
-#include "../internal.h"
+#include "../../internal.h"
 #include "internal.h"
