+++ crypto/fipsmodule/evp/p_rsa.c
@@ -65,10 +65,10 @@
 #include <openssl/nid.h>
 #include <openssl/rsa.h>
 
-#include "../internal.h"
-#include "../fipsmodule/rsa/internal.h"
+#include "../../internal.h"
+#include "../rsa/internal.h"
 #include "internal.h"
-#include "../rsa_extra/internal.h"
+#include "../../rsa_extra/internal.h"
 
 #define NO_PSS_SALT_LEN_RESTRICTION -1
