+++ crypto/fipsmodule/rsa/rsa.c
@@ -71,7 +71,7 @@
 #include <openssl/thread.h>
 
 #include "../../internal.h"
-#include "../../rsa_extra/internal.h"
+//#include "../../rsa_extra/internal.h"
 #include "../bn/internal.h"
 #include "../delocate.h"
 #include "internal.h"
